+++ NEWS.md
@@ -1,4 +1,4 @@
-#Rnets v 0.9.5
+#Rnets v 0.9.7
 
 ## Major Changes